feat: add refactorization on all paths and endpoinst for done date
- Add Task methods for verification of present data
- Add setters and getters on Task class
- Implement method to verify Task exists.
- Implement two methods for update doneDate on TaskDAO
- Add TaskDao Interface
- Refactorize TaskService to return  ResponseEntity
- Implement done and undone endpoints on ToDoController

+++ src/main/java/com/example/backend_breakable_toy_i_todoapp/service/TaskService.java
@@ -3,17 +3,19 @@ package com.example.backend_breakable_toy_i_todoapp.service;
 import com.example.backend_breakable_toy_i_todoapp.dao.TaskDAO;
 import com.example.backend_breakable_toy_i_todoapp.model.Task;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-// Repository must be injected (DAO)
 @Service
 public class TaskService {
     @Autowired
     private TaskDAO taskDAO;
+    public static List<String> priorities = List.of("high", "medium", "low");
 
     public List<Task> getAllTasks(){
         return new ArrayList<Task>(taskDAO.getAll().values());
@@ -23,14 +25,49 @@ public class TaskService {
         return  taskDAO.getTask(id);
     }
 
-    public void deleteTaskById(UUID id){
-        taskDAO.deleteTask(id);
+    public ResponseEntity<String> deleteTaskById(UUID id){
+        if(taskDAO.hasTask(id)){
+            taskDAO.deleteTask(id);
+            return new ResponseEntity<>("Task " + id.toString() + " was deleted!", HttpStatus.OK);
+        }
+        return  new ResponseEntity<>("Task not found", HttpStatus.NOT_FOUND);
     }
-    public void addTask(Task newTask){
-        taskDAO.addTask(newTask);
+    public  ResponseEntity<String> addTask(Task newTask){
+        if(newTask.hasRequiredFields()){
+            return  new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
+        }
+        if(!priorities.contains(newTask.getPriority())){
+            return  new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
+        }
+        return  new ResponseEntity<String>(taskDAO.addTask(newTask).toString(), HttpStatus.OK);
     }
-    public void updateTask(UUID id, Task updatedTask){
+    public ResponseEntity<String> updateTask(UUID id, Task updatedTask){
+        if(updatedTask.hasRequiredFields()){
+            return new ResponseEntity<>("Provided task has missing properties", HttpStatus.BAD_REQUEST);
+        }
+        if(!priorities.contains(updatedTask.getPriority())){
+            return  new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
+        }
+        if(!taskDAO.hasTask(id)){
+            return  new ResponseEntity<String>("Task not found", HttpStatus.BAD_REQUEST);
+        }
         taskDAO.updateTask(id, updatedTask);
+        return new ResponseEntity<String>("Task " + id.toString() + " updated.", HttpStatus.OK);
     }
-    //
+
+    public ResponseEntity<String> setDoneDateById(UUID id) {
+        if(!taskDAO.hasTask(id)){
+            return new ResponseEntity<>("Task not found for update", HttpStatus.BAD_REQUEST);
+        }
+        taskDAO.setDoneDate(id);
+        return new ResponseEntity<>("Task done date applied.", HttpStatus.OK);
+    }
+
+    public ResponseEntity<String> unsetDoneDateById(UUID id) {
+        if(!taskDAO.hasTask(id)){
+            return new ResponseEntity<>("Task not found for update", HttpStatus.BAD_REQUEST);
+        }
+        taskDAO.unsetDoneDate(id);
+        return new ResponseEntity<>("Task done date remove.", HttpStatus.OK);
+    }
 }
